+++ src/brod_kafka_request.erl
@@ -1,5 +1,5 @@
 %%%
-%%%   Copyright (c) 2017-2018 Klarna Bank AB (publ)
+%%%   Copyright (c) 2017-2021 Klarna Bank AB (publ)
 %%%
 %%%   Licensed under the Apache License, Version 2.0 (the "License");
 %%%   you may not use this file except in compliance with the License.
@@ -116,8 +116,8 @@ offset_fetch(Connection, GroupId, Topics0) ->
   Topics =
     lists:map(
       fun({Topic, Partitions}) ->
-        [ {topic, Topic}
-        , {partitions, [[{partition, P}] || P <- Partitions]}
+        [ {name, Topic}
+        , {partition_indexes, Partitions}
         ]
       end, Topics0),
   Body = [ {group_id, GroupId}