+++ test/brod_cg_commits_SUITE.erl
@@ -1,5 +1,5 @@
 %%%
-%%%   Copyright (c) 2017-2018, Klarna Bank AB (publ)
+%%%   Copyright (c) 2017-2021, Klarna Bank AB (publ)
 %%%
 %%%   Licensed under the Apache License, Version 2.0 (the "License");
 %%%   you may not use this file except in compliance with the License.
@@ -91,16 +91,16 @@ assert_offsets(ExpectedOffsets, Rsp) ->
 transform_rsp([], Acc) ->
   lists:keysort(1, Acc);
 transform_rsp([Struct | Rest], Acc) ->
-  Topic = kpro:find(topic, Struct),
-  PartitionRsp = kpro:find(partition_responses, Struct),
+  Topic = kpro:find(name, Struct),
+  PartitionRsp = kpro:find(partitions, Struct),
   Partitions = transform_rsp_partitions(PartitionRsp, []),
   transform_rsp(Rest, [{Topic, Partitions} | Acc]).
 
 transform_rsp_partitions([], Acc) ->
   lists:keysort(1, Acc);
 transform_rsp_partitions([Struct | Rest], Acc) ->
-  Partition = kpro:find(partition, Struct),
-  Offset = kpro:find(offset, Struct),
+  Partition = kpro:find(partition_index, Struct),
+  Offset = kpro:find(committed_offset, Struct),
   transform_rsp_partitions(Rest, [{Partition, Offset} | Acc]).
 
 do_commit(Topic, Offsets) ->